Fixed Task5.V1

+++ Tyuiu.VasilevAA.Sprint1.Task5.V1.Lib/DataService.cs
@@ -11,7 +11,7 @@ namespace Tyuiu.VasilevAA.Sprint1.Task5.V1.Lib
 {
     public class DataService : ISprint1Task5V1
     {
-        public double DistanceBetweenDots(double x1, double x2, double y1, double y2)
+        public int DistanceBetweenDots(double x1, double x2, double y1, double y2)
         {
             return Convert.ToInt32(Math.Sqrt(Math.Pow((x2 - x1), 2) + Math.Pow((y2 - y1), 2))); 
         }
